requirements.txt
bedrock-agentcore>=1.1.0
langchain>=1.2.0
langchain-groq>=1.1.0
langchain-huggingface>=1.2.0
langchain-community>=0.4.0
langgraph>=1.0.0
langgraph-prebuilt>=1.0.0
langgraph-checkpoint>=3.0.0
langgraph-checkpoint-aws>=1.0.0
faiss-cpu>=1.13.0
pydantic-settings>=2.0.0
groq>=1.0.0
python-dotenv>=1.0.0
